Fix type issue in Cookies. Add wait to Window maximize test to prevent failure when run too fast.

+++ lib/src/options.dart
@@ -64,7 +64,7 @@ class Cookie {
   factory Cookie.fromJson(Map<String, dynamic> json) {
     var expiry;
     if (json['expiry'] is num) {
-      expiry = new DateTime.fromMillisecondsSinceEpoch(json['expiry'] * 1000,
+      expiry = new DateTime.fromMillisecondsSinceEpoch(json['expiry'].toInt() * 1000,
           isUtc: true);
     }
     return new Cookie(json['name'], json['value'],
